==> Notebooks/1 - Enrichissement du dataset/3 - Nettoyage 1 du dataset.ipynb <==
import pandas as pd

df = pd.read_csv('Sources/final_dataset.csv')

df.head()

df = df.drop('Unnamed: 0', axis=1)

df.info()

df.columns.values

# Nettoyage du dataset

# 1ère liste de colonnes à supprimer car répétitives ou sans informations 
col_to_drop_1 = ['ATP', 'Winner', 'Loser', 'CBW', 'CBL', 'GBW', 'GBL', 'IWW', 'IWL', 
                 'SBW', 'SBL', 'B&WW', 'B&WL', 'UBW', 'UBL', 'LBW', 'LBL', 'SJW', 'SJL',
                 'MaxW', 'MaxL', 'AvgW', 'AvgL', 'city', 'winner_family_name', 'loser_family_name', 'tourney_year_id',
                 'tourney_order_x', 'tourney_name_x', 'tourney_slug_x', 'tourney_url_suffix_x', 'start_year', 'start_month',
                 'start_day', 'end_year', 'end_month', 'end_day', 'match_index', 'tourney_round_name',
                 'round_order', 'match_order', 'winner_slug_x', 'loser_slug_x', 'winner_seed', 'loser_seed', 
                 'match_score_tiebreaks', 'match_id', 'match_stats_url_suffix_x', 'tourney_order_y',
                 'tourney_type', 'tourney_name_y', 'tourney_id', 'tourney_slug_y','tourney_location', 'tourney_date', 
                 'year', 'tourney_month', 'tourney_day', 'tourney_singles_draw', 'tourney_doubles_draw','tourney_conditions', 
                 'tourney_surface', 'tourney_fin_commit_raw', 'tourney_url_suffix_y', 'singles_winner_name', 
                 'singles_winner_url', 'singles_winner_player_slug', 'singles_winner_player_id', 'doubles_winner_1_name', 
                 'doubles_winner_1_url', 'doubles_winner_1_player_slug', 'doubles_winner_1_player_id',
                 'doubles_winner_2_name', 'doubles_winner_2_url', 'doubles_winner_2_player_slug', 'doubles_winner_2_player_id',
                 'tourney_slug', 'match_stats_url_suffix_y', 'match_time', 'year1', 'winner_player_slug', 'winner_first_name',
                 'winner_last_name', 'winner_player_url', 'winner_residence', 'winner_birth_year',
                 'winner_birth_month', 'winner_birth_day', 'winner_weight_lbs', 'winner_height_ft', 'winner_height_inches', 
                 'loser_player_slug', 'loser_first_name', 'loser_last_name', 'loser_player_url', 'loser_residence',
                 'loser_birth_year', 'loser_birth_month', 'loser_birth_day', 'loser_weight_lbs', 'loser_height_ft', 'loser_height_inches']

df = df.drop(col_to_drop_1, axis=1)

df.info()

# Comparaison des valeurs uniques pour le winner
print(len(df['winner_name'].unique())) # col to keep
print(len(df['winner_player_id'].unique())) # col to keep


# Comparaison des valeurs uniques pour le loser
print(len(df['loser_name'].unique())) # col to keep
print(len(df['loser_player_id'].unique())) # col to keep


# recherche et modification du loser_name ayant 2 id
loser_name = df.groupby('loser_player_id').agg({'loser_name':'unique'})
liste = []
for value in loser_name['loser_name']:
    if len(value) > 1:
        liste.append(value)
print(liste)

df['loser_name'] = df['loser_name'].replace(liste[0][1], liste[0][0])

print(len(df['loser_name'].unique())) # col to keep
print(len(df['loser_player_id'].unique())) # col to keep

# changement de type pour WRank et LRank + remplacement des nan par 0
df['WRank'] = df['WRank'].fillna(0).astype('int')

df['LRank'] = df['LRank'].fillna(0).replace('NR', 0).astype('float')

df['LRank'] = df['LRank'].astype('int')

df.info()

df['Series'].value_counts()

# suppression de 'winner_slug_y' et 'loser_slug_y'
df = df.drop(['winner_slug_y', 'loser_slug_y'], axis=1)

df.info()

# comparaison prix et currency
df[['currency_x', 'prize_money', 'currency_y', 'tourney_fin_commit']]

# suppression de 'currency_x' et 'prize_money'
df = df.drop(['currency_x', 'prize_money'], axis=1)

df.info()

df['Comment'].value_counts()

df.to_csv('Sources/final_dataset_firstclean.csv')

df.info()
