==> Notebooks/1 - Enrichissement du dataset/6 - Preprocessing du final_dataset_thirdclean to rolling with bookmakers.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

df = pd.read_csv('Sources/final_dataset_thirdclean.csv')

df.info()

df.rename({'B365W':'winner_B365', 'B365L':'loser_B365', 'PSW':'winner_PS', 'PSL':'loser_PS'}, axis=1, inplace=True)

df.info()

# df_winner
df_winner = df[['Tournament', 'Date', 'year', 'month', 'day', 'Series', 'Court',
       'Surface', 'prize_money_euro', 'Round', 'Best of', 'winner_B365', 'winner_PS',
       'match_duration', 'nb_sets', 'nb_games', 'nb_tiebreaks',
       'nb_points', 'winner_name', 'winner_rank', 'winner_height_cm',
       'winner_weight_kg', 'winner_handedness', 'winner_backhand',
       'winner_at_home', 'winner_age', 'winner_nb_year_xp',
       'winner_sets_won', 'winner_games_won', 'winner_tiebreaks_won',
       'winner_points_won', 'winner_serve_rating', 'winner_aces',
       'winner_double_faults', 'winner_first_serves_in',
       'winner_first_serves_total', 'winner_first_serve_points_won',
       'winner_first_serve_points_total',
       'winner_second_serve_points_won',
       'winner_second_serve_points_total', 'winner_break_points_saved',
       'winner_break_points_serve_total', 'winner_service_games_played',
       'winner_return_rating', 'winner_first_serve_return_won',
       'winner_first_serve_return_total',
       'winner_second_serve_return_won',
       'winner_second_serve_return_total',
       'winner_break_points_converted',
       'winner_break_points_return_total', 'winner_return_games_played',
       'winner_service_points_won', 'winner_service_points_total',
       'winner_return_points_won', 'winner_return_points_total',
       'winner_under_pressure_rating']]

df_winner['win_rate'] = 1

# df_loser
df_loser = df[['Tournament', 'Date', 'year', 'month', 'day', 'Series', 'Court',
       'Surface', 'prize_money_euro', 'Round', 'Best of', 'loser_B365', 'loser_PS',
       'match_duration', 'nb_sets', 'nb_games', 'nb_tiebreaks',
       'nb_points', 'loser_name', 'loser_rank',
       'loser_height_cm', 'loser_weight_kg', 'loser_handedness',
       'loser_backhand', 'loser_at_home', 'loser_age', 'loser_nb_year_xp',
       'loser_sets_won', 'loser_games_won', 'loser_tiebreaks_won',
       'loser_points_won', 'loser_serve_rating', 'loser_aces',
       'loser_double_faults', 'loser_first_serves_in',
       'loser_first_serves_total', 'loser_first_serve_points_won',
       'loser_first_serve_points_total', 'loser_second_serve_points_won',
       'loser_second_serve_points_total', 'loser_break_points_saved',
       'loser_break_points_serve_total', 'loser_service_games_played',
       'loser_return_rating', 'loser_first_serve_return_won',
       'loser_first_serve_return_total', 'loser_second_serve_return_won',
       'loser_second_serve_return_total', 'loser_break_points_converted',
       'loser_break_points_return_total', 'loser_return_games_played',
       'loser_service_points_won', 'loser_service_points_total',
       'loser_return_points_won', 'loser_return_points_total',
       'loser_under_pressure_rating']]

df_loser['win_rate'] = 0

col_names = ['Tournament', 'Date', 'year', 'month', 'day', 'Series', 'Court',
             'Surface', 'prize_money_euro', 'Round', 'Best of', 'B365', 'PS',
             'match_duration', 'nb_sets', 'nb_games', 'nb_tiebreaks',
             'nb_points', 'name', 'rank',
       'height_cm', 'weight_kg', 'handedness',
       'backhand', 'at_home', 'age', 'nb_year_xp',
       'sets_won', 'games_won', 'tiebreaks_won',
       'points_won', 'serve_rating', 'aces',
       'double_faults', 'first_serves_in',
       'first_serves_total', 'first_serve_points_won',
       'first_serve_points_total', 'second_serve_points_won',
       'second_serve_points_total', 'break_points_saved',
       'break_points_serve_total', 'service_games_played',
       'return_rating', 'first_serve_return_won',
       'first_serve_return_total', 'second_serve_return_won',
       'second_serve_return_total', 'break_points_converted',
       'break_points_return_total', 'return_games_played',
       'service_points_won', 'service_points_total',
       'return_points_won', 'return_points_total',
       'under_pressure_rating', 'win_rate']

df_winner.columns = col_names
df_loser.columns = col_names

# concatenation
df_new = pd.concat([df_winner,df_loser], axis=0)

df_total = df_new.sort_values(by=['name', 'Date'])

df_test = df_total[['name', 'match_duration', 'nb_sets', 'nb_games', 'nb_tiebreaks',
             'nb_points', 'sets_won', 'games_won', 'tiebreaks_won', 'points_won',
       'serve_rating', 'aces', 'double_faults', 'first_serves_in',
       'first_serves_total', 'first_serve_points_won',
       'first_serve_points_total', 'second_serve_points_won',
       'second_serve_points_total', 'break_points_saved',
       'break_points_serve_total', 'service_games_played',
       'return_rating', 'first_serve_return_won',
       'first_serve_return_total', 'second_serve_return_won',
       'second_serve_return_total', 'break_points_converted',
       'break_points_return_total', 'return_games_played',
       'service_points_won', 'service_points_total', 'return_points_won',
       'return_points_total', 'under_pressure_rating', 'win_rate']]

df_rolling_30 = df_test.groupby('name').rolling(window = 30, min_periods = 0, center=False).mean()

tableau = np.array(df_rolling_30)

tableau.shape

tableau

#insertion ligne de 0
zeros = np.zeros((1,tableau.shape[1]))

#concatenation
new_tableau = np.concatenate([zeros, tableau], axis=0)

new_tableau = np.delete(new_tableau, -1, axis=0)

new_tableau.shape

condition = []
names = list(df_test['name'])
for index, name in enumerate(names):
    if name != names[index - 1]:
        condition.append(0)                           
    else:
        condition.append(1)
condition_array = np.array(condition)

condition_array = condition_array.reshape((condition_array.shape[0], 1))

condition_array

df_new_2 = pd.DataFrame(new_tableau * condition_array, index=df_test.index, columns=df_rolling_30.columns)
df_new_2['name'] = df_test['name']

#On réinsère la colonne win pour reconstruire le df de base
df_new_2['win'] = df_test['win_rate']

df_commun = df_total[['Tournament', 'Date', 'year', 'month', 'day', 'Series', 'Court',
       'Surface', 'prize_money_euro', 'Round', 'Best of', 'B365', 'PS', 'rank', 'height_cm', 'weight_kg',
       'handedness', 'backhand', 'at_home', 'age', 'nb_year_xp']]

df_new_3 = pd.concat([df_commun, df_new_2], axis=1)

df_winner_new = df_new_3[df_new_3['win'] == 1]
df_loser_new = df_new_3[df_new_3['win'] == 0]

df_winner_new = df_winner_new.sort_index()

df_loser_new = df_loser_new.sort_index()

df_loser_to_concat = df_loser_new[['B365', 'PS', 'name', 'rank',
       'height_cm', 'weight_kg', 'handedness', 'backhand', 'at_home',
       'age', 'nb_year_xp', 'match_duration', 'nb_sets', 'nb_games',
       'nb_tiebreaks', 'nb_points', 'sets_won', 'games_won',
       'tiebreaks_won', 'points_won', 'serve_rating', 'aces',
       'double_faults', 'first_serves_in', 'first_serves_total',
       'first_serve_points_won', 'first_serve_points_total',
       'second_serve_points_won', 'second_serve_points_total',
       'break_points_saved', 'break_points_serve_total',
       'service_games_played', 'return_rating', 'first_serve_return_won',
       'first_serve_return_total', 'second_serve_return_won',
       'second_serve_return_total', 'break_points_converted',
       'break_points_return_total', 'return_games_played',
       'service_points_won', 'service_points_total', 'return_points_won',
       'return_points_total', 'under_pressure_rating', 'win_rate']]

loser_col_names = ['loser_B365', 'loser_PS', 'loser_name', 'loser_rank',
       'loser_height_cm', 'loser_weight_kg', 'loser_handedness', 'loser_backhand', 'loser_at_home',
       'loser_age', 'loser_nb_year_xp', 'loser_match_duration', 'loser_nb_sets', 'loser_nb_games',
       'loser_nb_tiebreaks', 'loser_nb_points', 'loser_sets_won', 'loser_games_won',
       'loser_tiebreaks_won', 'loser_points_won', 'loser_serve_rating', 'loser_aces',
       'loser_double_faults', 'loser_first_serves_in', 'loser_first_serves_total',
       'loser_first_serve_points_won', 'loser_first_serve_points_total',
       'loser_second_serve_points_won', 'loser_second_serve_points_total',
       'loser_break_points_saved', 'loser_break_points_serve_total',
       'loser_service_games_played', 'loser_return_rating', 'loser_first_serve_return_won',
       'loser_first_serve_return_total', 'loser_second_serve_return_won',
       'loser_second_serve_return_total', 'loser_break_points_converted',
       'loser_break_points_return_total', 'loser_return_games_played',
       'loser_service_points_won', 'loser_service_points_total', 'loser_return_points_won',
       'loser_return_points_total', 'loser_under_pressure_rating', 'loser_win_rate']

df_loser_to_concat.columns = loser_col_names

df_winner_to_concat = df_winner_new[['Tournament', 'Date', 'year', 'month', 'day', 'Series', 'Court',
       'Surface', 'prize_money_euro', 'Round', 'Best of', 'B365', 'PS', 'name', 'rank',
       'height_cm', 'weight_kg', 'handedness', 'backhand', 'at_home',
       'age', 'nb_year_xp', 'match_duration', 'nb_sets', 'nb_games',
       'nb_tiebreaks', 'nb_points', 'sets_won', 'games_won',
       'tiebreaks_won', 'points_won', 'serve_rating', 'aces',
       'double_faults', 'first_serves_in', 'first_serves_total',
       'first_serve_points_won', 'first_serve_points_total',
       'second_serve_points_won', 'second_serve_points_total',
       'break_points_saved', 'break_points_serve_total',
       'service_games_played', 'return_rating', 'first_serve_return_won',
       'first_serve_return_total', 'second_serve_return_won',
       'second_serve_return_total', 'break_points_converted',
       'break_points_return_total', 'return_games_played',
       'service_points_won', 'service_points_total', 'return_points_won',
       'return_points_total', 'under_pressure_rating', 'win_rate']]

winner_col_names = ['Tournament', 'Date', 'year', 'month', 'day', 'Series', 'Court',
       'Surface', 'prize_money_euro', 'Round', 'Best of', 'winner_B365', 'winner_PS', 'winner_name', 'winner_rank',
       'winner_height_cm', 'winner_weight_kg', 'winner_handedness', 'winner_backhand', 'winner_at_home',
       'winner_age', 'winner_nb_year_xp', 'winner_match_duration', 'winner_nb_sets', 'winner_nb_games',
       'winner_nb_tiebreaks', 'winner_nb_points', 'winner_sets_won', 'winner_games_won',
       'winner_tiebreaks_won', 'winner_points_won', 'winner_serve_rating', 'winner_aces',
       'winner_double_faults', 'winner_first_serves_in', 'winner_first_serves_total',
       'winner_first_serve_points_won', 'winner_first_serve_points_total',
       'winner_second_serve_points_won', 'winner_second_serve_points_total',
       'winner_break_points_saved', 'winner_break_points_serve_total',
       'winner_service_games_played', 'winner_return_rating', 'winner_first_serve_return_won',
       'winner_first_serve_return_total', 'winner_second_serve_return_won',
       'winner_second_serve_return_total', 'winner_break_points_converted',
       'winner_break_points_return_total', 'winner_return_games_played',
       'winner_service_points_won', 'winner_service_points_total', 'winner_return_points_won',
       'winner_return_points_total', 'winner_under_pressure_rating', 'winner_win_rate']

df_winner_to_concat.columns = winner_col_names

df_final_rolling_30 = pd.concat([df_winner_to_concat, df_loser_to_concat], axis=1)

df_final_rolling_30.info()

df = df_final_rolling_30

player_1_B365 = []
player_1_PS = []
player_1_name = []
player_1_rank = []
player_1_height_cm = []
player_1_weight_kg = []
player_1_handedness = []
player_1_backhand = []
player_1_at_home = []
player_1_age = []
player_1_nb_year_xp = []
player_1_match_duration = []
player_1_nb_sets = []
player_1_nb_games = []
player_1_nb_tiebreaks = []
player_1_nb_points = []
player_1_sets_won = []
player_1_games_won = []
player_1_tiebreaks_won = []
player_1_points_won = []
player_1_serve_rating = []
player_1_aces = []
player_1_double_faults = []
player_1_first_serves_in = []
player_1_first_serves_total = []
player_1_first_serve_points_won = []
player_1_first_serve_points_total = []
player_1_second_serve_points_won = []
player_1_second_serve_points_total = []
player_1_break_points_saved = []
player_1_break_points_serve_total = []
player_1_service_games_played = []
player_1_return_rating = []
player_1_first_serve_return_won = []
player_1_first_serve_return_total = []
player_1_second_serve_return_won = []
player_1_second_serve_return_total = []
player_1_break_points_converted = []
player_1_break_points_return_total = []
player_1_return_games_played = []
player_1_service_points_won = []
player_1_service_points_total = []
player_1_return_points_won = []
player_1_return_points_total = []
player_1_under_pressure_rating = []
player_1_win_rate = []
player_1_win = []

player_2_B365 = []
player_2_PS = []
player_2_name = []
player_2_rank = []
player_2_height_cm = []
player_2_weight_kg = []
player_2_handedness = []
player_2_backhand = []
player_2_at_home = []
player_2_age = []
player_2_nb_year_xp = []
player_2_match_duration = []
player_2_nb_sets = []
player_2_nb_games = []
player_2_nb_tiebreaks = []
player_2_nb_points = []
player_2_sets_won = []
player_2_games_won = []
player_2_tiebreaks_won = []
player_2_points_won = []
player_2_serve_rating = []
player_2_aces = []
player_2_double_faults = []
player_2_first_serves_in = []
player_2_first_serves_total = []
player_2_first_serve_points_won = []
player_2_first_serve_points_total = []
player_2_second_serve_points_won = []
player_2_second_serve_points_total = []
player_2_break_points_saved = []
player_2_break_points_serve_total = []
player_2_service_games_played = []
player_2_return_rating = []
player_2_first_serve_return_won = []
player_2_first_serve_return_total = []
player_2_second_serve_return_won = []
player_2_second_serve_return_total = []
player_2_break_points_converted = []
player_2_break_points_return_total = []
player_2_return_games_played = []
player_2_service_points_won = []
player_2_service_points_total = []
player_2_return_points_won = []
player_2_return_points_total = []
player_2_under_pressure_rating = []
player_2_win_rate = []

np.random.seed(1)
for winner_B365, winner_PS, winner_name, winner_rank, winner_height_cm, winner_weight_kg, winner_handedness, winner_backhand, winner_at_home, winner_age, winner_nb_year_xp, winner_match_duration, winner_nb_sets, winner_nb_games, winner_nb_tiebreaks, winner_nb_points, winner_sets_won, winner_games_won, winner_tiebreaks_won, winner_points_won, winner_serve_rating, winner_aces, winner_double_faults, winner_first_serves_in, winner_first_serves_total, winner_first_serve_points_won, winner_first_serve_points_total, winner_second_serve_points_won, winner_second_serve_points_total, winner_break_points_saved, winner_break_points_serve_total, winner_service_games_played, winner_return_rating, winner_first_serve_return_won, winner_first_serve_return_total, winner_second_serve_return_won, winner_second_serve_return_total, winner_break_points_converted, winner_break_points_return_total, winner_return_games_played, winner_service_points_won, winner_service_points_total, winner_return_points_won, winner_return_points_total, winner_under_pressure_rating, winner_win_rate, loser_B365, loser_PS, loser_name, loser_rank, loser_height_cm, loser_weight_kg, loser_handedness, loser_backhand, loser_at_home, loser_age, loser_nb_year_xp, loser_match_duration, loser_nb_sets, loser_nb_games, loser_nb_tiebreaks, loser_nb_points, loser_sets_won, loser_games_won, loser_tiebreaks_won, loser_points_won, loser_serve_rating, loser_aces, loser_double_faults, loser_first_serves_in, loser_first_serves_total, loser_first_serve_points_won, loser_first_serve_points_total, loser_second_serve_points_won, loser_second_serve_points_total, loser_break_points_saved, loser_break_points_serve_total, loser_service_games_played, loser_return_rating, loser_first_serve_return_won, loser_first_serve_return_total, loser_second_serve_return_won, loser_second_serve_return_total, loser_break_points_converted, loser_break_points_return_total, loser_return_games_played, loser_service_points_won, loser_service_points_total, loser_return_points_won, loser_return_points_total, loser_under_pressure_rating, loser_win_rate in zip(df['winner_B365'], df['winner_PS'], df['winner_name'],df['winner_rank'],df['winner_height_cm'],df['winner_weight_kg'],df['winner_handedness'],df['winner_backhand'],df['winner_at_home'],df['winner_age'],df['winner_nb_year_xp'],df['winner_match_duration'],df['winner_nb_sets'],df['winner_nb_games'],df['winner_nb_tiebreaks'],df['winner_nb_points'],df['winner_sets_won'],df['winner_games_won'],df['winner_tiebreaks_won'],df['winner_points_won'],df['winner_serve_rating'],df['winner_aces'],df['winner_double_faults'],df['winner_first_serves_in'],df['winner_first_serves_total'],df['winner_first_serve_points_won'],df['winner_first_serve_points_total'],df['winner_second_serve_points_won'],df['winner_second_serve_points_total'],df['winner_break_points_saved'],df['winner_break_points_serve_total'],df['winner_service_games_played'],df['winner_return_rating'],df['winner_first_serve_return_won'],df['winner_first_serve_return_total'],df['winner_second_serve_return_won'],df['winner_second_serve_return_total'],df['winner_break_points_converted'],df['winner_break_points_return_total'],df['winner_return_games_played'],df['winner_service_points_won'],df['winner_service_points_total'],df['winner_return_points_won'],df['winner_return_points_total'],df['winner_under_pressure_rating'],df['winner_win_rate'], df['loser_B365'], df['loser_PS'], df['loser_name'], df['loser_rank'], df['loser_height_cm'], df['loser_weight_kg'], df['loser_handedness'], df['loser_backhand'], df['loser_at_home'], df['loser_age'], df['loser_nb_year_xp'], df['loser_match_duration'], df['loser_nb_sets'], df['loser_nb_games'], df['loser_nb_tiebreaks'], df['loser_nb_points'], df['loser_sets_won'], df['loser_games_won'], df['loser_tiebreaks_won'], df['loser_points_won'], df['loser_serve_rating'], df['loser_aces'], df['loser_double_faults'], df['loser_first_serves_in'], df['loser_first_serves_total'], df['loser_first_serve_points_won'], df['loser_first_serve_points_total'], df['loser_second_serve_points_won'], df['loser_second_serve_points_total'], df['loser_break_points_saved'], df['loser_break_points_serve_total'], df['loser_service_games_played'], df['loser_return_rating'], df['loser_first_serve_return_won'], df['loser_first_serve_return_total'], df['loser_second_serve_return_won'], df['loser_second_serve_return_total'], df['loser_break_points_converted'], df['loser_break_points_return_total'], df['loser_return_games_played'], df['loser_service_points_won'], df['loser_service_points_total'], df['loser_return_points_won'], df['loser_return_points_total'], df['loser_under_pressure_rating'], df['loser_win_rate']):
    num_winner = np.random.choice(np.array([1,2]))
    if num_winner == 1:
        player_1_B365.append(winner_B365)
        player_1_PS.append(winner_PS)
        player_1_name.append(winner_name)
        player_1_rank.append(winner_rank)
        player_1_height_cm.append(winner_height_cm)
        player_1_weight_kg.append(winner_weight_kg)
        player_1_handedness.append(winner_handedness)
        player_1_backhand.append(winner_backhand)
        player_1_at_home.append(winner_at_home)
        player_1_age.append(winner_age)
        player_1_nb_year_xp.append(winner_nb_year_xp)
        player_1_match_duration.append(winner_match_duration)
        player_1_nb_sets.append(winner_nb_sets)
        player_1_nb_games.append(winner_nb_games)
        player_1_nb_tiebreaks.append(winner_nb_tiebreaks)
        player_1_nb_points.append(winner_nb_points)
        player_1_sets_won.append(winner_sets_won)
        player_1_games_won.append(winner_games_won)
        player_1_tiebreaks_won.append(winner_tiebreaks_won)
        player_1_points_won.append(winner_points_won)
        player_1_serve_rating.append(winner_serve_rating)
        player_1_aces.append(winner_aces)
        player_1_double_faults.append(winner_double_faults)
        player_1_first_serves_in.append(winner_first_serves_in)
        player_1_first_serves_total.append(winner_first_serves_total)
        player_1_first_serve_points_won.append(winner_first_serve_points_won)
        player_1_first_serve_points_total.append(winner_first_serve_points_total)
        player_1_second_serve_points_won.append(winner_second_serve_points_won)
        player_1_second_serve_points_total.append(winner_second_serve_points_total)
        player_1_break_points_saved.append(winner_break_points_saved)
        player_1_break_points_serve_total.append(winner_break_points_serve_total)
        player_1_service_games_played.append(winner_service_games_played)
        player_1_return_rating.append(winner_return_rating)
        player_1_first_serve_return_won.append(winner_first_serve_return_won)
        player_1_first_serve_return_total.append(winner_first_serve_return_total)
        player_1_second_serve_return_won.append(winner_second_serve_return_won)
        player_1_second_serve_return_total.append(winner_second_serve_return_total)
        player_1_break_points_converted.append(winner_break_points_converted)
        player_1_break_points_return_total.append(winner_break_points_return_total)
        player_1_return_games_played.append(winner_return_games_played)
        player_1_service_points_won.append(winner_service_points_won)
        player_1_service_points_total.append(winner_service_points_total)
        player_1_return_points_won.append(winner_return_points_won)
        player_1_return_points_total.append(winner_return_points_total)
        player_1_under_pressure_rating.append(winner_under_pressure_rating)
        player_1_win_rate.append(winner_win_rate)
        player_1_win.append(1)
        player_2_B365.append(loser_B365)
        player_2_PS.append(loser_PS)
        player_2_name.append(loser_name)
        player_2_rank.append(loser_rank)
        player_2_height_cm.append(loser_height_cm)
        player_2_weight_kg.append(loser_weight_kg)
        player_2_handedness.append(loser_handedness)
        player_2_backhand.append(loser_backhand)
        player_2_at_home.append(loser_at_home)
        player_2_age.append(loser_age)
        player_2_nb_year_xp.append(loser_nb_year_xp)
        player_2_match_duration.append(loser_match_duration)
        player_2_nb_sets.append(loser_nb_sets)
        player_2_nb_games.append(loser_nb_games)
        player_2_nb_tiebreaks.append(loser_nb_tiebreaks)
        player_2_nb_points.append(loser_nb_points)
        player_2_sets_won.append(loser_sets_won)
        player_2_games_won.append(loser_games_won)
        player_2_tiebreaks_won.append(loser_tiebreaks_won)
        player_2_points_won.append(loser_points_won)
        player_2_serve_rating.append(loser_serve_rating)
        player_2_aces.append(loser_aces)
        player_2_double_faults.append(loser_double_faults)
        player_2_first_serves_in.append(loser_first_serves_in)
        player_2_first_serves_total.append(loser_first_serves_total)
        player_2_first_serve_points_won.append(loser_first_serve_points_won)
        player_2_first_serve_points_total.append(loser_first_serve_points_total)
        player_2_second_serve_points_won.append(loser_second_serve_points_won)
        player_2_second_serve_points_total.append(loser_second_serve_points_total)
        player_2_break_points_saved.append(loser_break_points_saved)
        player_2_break_points_serve_total.append(loser_break_points_serve_total)
        player_2_service_games_played.append(loser_service_games_played)
        player_2_return_rating.append(loser_return_rating)
        player_2_first_serve_return_won.append(loser_first_serve_return_won)
        player_2_first_serve_return_total.append(loser_first_serve_return_total)
        player_2_second_serve_return_won.append(loser_second_serve_return_won)
        player_2_second_serve_return_total.append(loser_second_serve_return_total)
        player_2_break_points_converted.append(loser_break_points_converted)
        player_2_break_points_return_total.append(loser_break_points_return_total)
        player_2_return_games_played.append(loser_return_games_played)
        player_2_service_points_won.append(loser_service_points_won)
        player_2_service_points_total.append(loser_service_points_total)
        player_2_return_points_won.append(loser_return_points_won)
        player_2_return_points_total.append(loser_return_points_total)
        player_2_under_pressure_rating.append(loser_under_pressure_rating)
        player_2_win_rate.append(loser_win_rate)
    else:
        player_2_B365.append(winner_B365)
        player_2_PS.append(winner_PS)
        player_2_name.append(winner_name)
        player_2_rank.append(winner_rank)
        player_2_height_cm.append(winner_height_cm)
        player_2_weight_kg.append(winner_weight_kg)
        player_2_handedness.append(winner_handedness)
        player_2_backhand.append(winner_backhand)
        player_2_at_home.append(winner_at_home)
        player_2_age.append(winner_age)
        player_2_nb_year_xp.append(winner_nb_year_xp)
        player_2_match_duration.append(winner_match_duration)
        player_2_nb_sets.append(winner_nb_sets)
        player_2_nb_games.append(winner_nb_games)
        player_2_nb_tiebreaks.append(winner_nb_tiebreaks)
        player_2_nb_points.append(winner_nb_points)
        player_2_sets_won.append(winner_sets_won)
        player_2_games_won.append(winner_games_won)
        player_2_tiebreaks_won.append(winner_tiebreaks_won)
        player_2_points_won.append(winner_points_won)
        player_2_serve_rating.append(winner_serve_rating)
        player_2_aces.append(winner_aces)
        player_2_double_faults.append(winner_double_faults)
        player_2_first_serves_in.append(winner_first_serves_in)
        player_2_first_serves_total.append(winner_first_serves_total)
        player_2_first_serve_points_won.append(winner_first_serve_points_won)
        player_2_first_serve_points_total.append(winner_first_serve_points_total)
        player_2_second_serve_points_won.append(winner_second_serve_points_won)
        player_2_second_serve_points_total.append(winner_second_serve_points_total)
        player_2_break_points_saved.append(winner_break_points_saved)
        player_2_break_points_serve_total.append(winner_break_points_serve_total)
        player_2_service_games_played.append(winner_service_games_played)
        player_2_return_rating.append(winner_return_rating)
        player_2_first_serve_return_won.append(winner_first_serve_return_won)
        player_2_first_serve_return_total.append(winner_first_serve_return_total)
        player_2_second_serve_return_won.append(winner_second_serve_return_won)
        player_2_second_serve_return_total.append(winner_second_serve_return_total)
        player_2_break_points_converted.append(winner_break_points_converted)
        player_2_break_points_return_total.append(winner_break_points_return_total)
        player_2_return_games_played.append(winner_return_games_played)
        player_2_service_points_won.append(winner_service_points_won)
        player_2_service_points_total.append(winner_service_points_total)
        player_2_return_points_won.append(winner_return_points_won)
        player_2_return_points_total.append(winner_return_points_total)
        player_2_under_pressure_rating.append(winner_under_pressure_rating)
        player_2_win_rate.append(winner_win_rate)
        player_1_win.append(0)
        player_1_B365.append(loser_B365)
        player_1_PS.append(loser_PS)
        player_1_name.append(loser_name)
        player_1_rank.append(loser_rank)
        player_1_height_cm.append(loser_height_cm)
        player_1_weight_kg.append(loser_weight_kg)
        player_1_handedness.append(loser_handedness)
        player_1_backhand.append(loser_backhand)
        player_1_at_home.append(loser_at_home)
        player_1_age.append(loser_age)
        player_1_nb_year_xp.append(loser_nb_year_xp)
        player_1_match_duration.append(loser_match_duration)
        player_1_nb_sets.append(loser_nb_sets)
        player_1_nb_games.append(loser_nb_games)
        player_1_nb_tiebreaks.append(loser_nb_tiebreaks)
        player_1_nb_points.append(loser_nb_points)
        player_1_sets_won.append(loser_sets_won)
        player_1_games_won.append(loser_games_won)
        player_1_tiebreaks_won.append(loser_tiebreaks_won)
        player_1_points_won.append(loser_points_won)
        player_1_serve_rating.append(loser_serve_rating)
        player_1_aces.append(loser_aces)
        player_1_double_faults.append(loser_double_faults)
        player_1_first_serves_in.append(loser_first_serves_in)
        player_1_first_serves_total.append(loser_first_serves_total)
        player_1_first_serve_points_won.append(loser_first_serve_points_won)
        player_1_first_serve_points_total.append(loser_first_serve_points_total)
        player_1_second_serve_points_won.append(loser_second_serve_points_won)
        player_1_second_serve_points_total.append(loser_second_serve_points_total)
        player_1_break_points_saved.append(loser_break_points_saved)
        player_1_break_points_serve_total.append(loser_break_points_serve_total)
        player_1_service_games_played.append(loser_service_games_played)
        player_1_return_rating.append(loser_return_rating)
        player_1_first_serve_return_won.append(loser_first_serve_return_won)
        player_1_first_serve_return_total.append(loser_first_serve_return_total)
        player_1_second_serve_return_won.append(loser_second_serve_return_won)
        player_1_second_serve_return_total.append(loser_second_serve_return_total)
        player_1_break_points_converted.append(loser_break_points_converted)
        player_1_break_points_return_total.append(loser_break_points_return_total)
        player_1_return_games_played.append(loser_return_games_played)
        player_1_service_points_won.append(loser_service_points_won)
        player_1_service_points_total.append(loser_service_points_total)
        player_1_return_points_won.append(loser_return_points_won)
        player_1_return_points_total.append(loser_return_points_total)
        player_1_under_pressure_rating.append(loser_under_pressure_rating)
        player_1_win_rate.append(loser_win_rate)

liste_1 = ['player_1_B365',
           'player_1_PS',
           'player_1_name',
'player_1_rank',
'player_1_height_cm',
'player_1_weight_kg',
'player_1_handedness',
'player_1_backhand',
'player_1_at_home',
'player_1_age',
'player_1_nb_year_xp',
'player_1_match_duration',
'player_1_nb_sets',
'player_1_nb_games',
'player_1_nb_tiebreaks',
'player_1_nb_points',
'player_1_sets_won',
'player_1_games_won',
'player_1_tiebreaks_won',
'player_1_points_won',
'player_1_serve_rating',
'player_1_aces',
'player_1_double_faults',
'player_1_first_serves_in',
'player_1_first_serves_total',
'player_1_first_serve_points_won',
'player_1_first_serve_points_total',
'player_1_second_serve_points_won',
'player_1_second_serve_points_total',
'player_1_break_points_saved',
'player_1_break_points_serve_total',
'player_1_service_games_played',
'player_1_return_rating',
'player_1_first_serve_return_won',
'player_1_first_serve_return_total',
'player_1_second_serve_return_won',
'player_1_second_serve_return_total',
'player_1_break_points_converted',
'player_1_break_points_return_total',
'player_1_return_games_played',
'player_1_service_points_won',
'player_1_service_points_total',
'player_1_return_points_won',
'player_1_return_points_total',
'player_1_under_pressure_rating',
'player_1_win_rate',
'player_1_win',
           'player_2_B365',
           'player_2_PS',
'player_2_name',
'player_2_rank',
'player_2_height_cm',
'player_2_weight_kg',
'player_2_handedness',
'player_2_backhand',
'player_2_at_home',
'player_2_age',
'player_2_nb_year_xp',
'player_2_match_duration',
'player_2_nb_sets',
'player_2_nb_games',
'player_2_nb_tiebreaks',
'player_2_nb_points',
'player_2_sets_won',
'player_2_games_won',
'player_2_tiebreaks_won',
'player_2_points_won',
'player_2_serve_rating',
'player_2_aces',
'player_2_double_faults',
'player_2_first_serves_in',
'player_2_first_serves_total',
'player_2_first_serve_points_won',
'player_2_first_serve_points_total',
'player_2_second_serve_points_won',
'player_2_second_serve_points_total',
'player_2_break_points_saved',
'player_2_break_points_serve_total',
'player_2_service_games_played',
'player_2_return_rating',
'player_2_first_serve_return_won',
'player_2_first_serve_return_total',
'player_2_second_serve_return_won',
'player_2_second_serve_return_total',
'player_2_break_points_converted',
'player_2_break_points_return_total',
'player_2_return_games_played',
'player_2_service_points_won',
'player_2_service_points_total',
'player_2_return_points_won',
'player_2_return_points_total',
'player_2_under_pressure_rating',
'player_2_win_rate']

liste_2 = [player_1_B365,
           player_1_PS,
           player_1_name,
player_1_rank,
player_1_height_cm,
player_1_weight_kg,
player_1_handedness,
player_1_backhand,
player_1_at_home,
player_1_age,
player_1_nb_year_xp,
player_1_match_duration,
player_1_nb_sets,
player_1_nb_games,
player_1_nb_tiebreaks,
player_1_nb_points,
player_1_sets_won,
player_1_games_won,
player_1_tiebreaks_won,
player_1_points_won,
player_1_serve_rating,
player_1_aces,
player_1_double_faults,
player_1_first_serves_in,
player_1_first_serves_total,
player_1_first_serve_points_won,
player_1_first_serve_points_total,
player_1_second_serve_points_won,
player_1_second_serve_points_total,
player_1_break_points_saved,
player_1_break_points_serve_total,
player_1_service_games_played,
player_1_return_rating,
player_1_first_serve_return_won,
player_1_first_serve_return_total,
player_1_second_serve_return_won,
player_1_second_serve_return_total,
player_1_break_points_converted,
player_1_break_points_return_total,
player_1_return_games_played,
player_1_service_points_won,
player_1_service_points_total,
player_1_return_points_won,
player_1_return_points_total,
player_1_under_pressure_rating,
player_1_win_rate,
player_1_win,
           player_2_B365,
           player_2_PS,
player_2_name,
player_2_rank,
player_2_height_cm,
player_2_weight_kg,
player_2_handedness,
player_2_backhand,
player_2_at_home,
player_2_age,
player_2_nb_year_xp,
player_2_match_duration,
player_2_nb_sets,
player_2_nb_games,
player_2_nb_tiebreaks,
player_2_nb_points,
player_2_sets_won,
player_2_games_won,
player_2_tiebreaks_won,
player_2_points_won,
player_2_serve_rating,
player_2_aces,
player_2_double_faults,
player_2_first_serves_in,
player_2_first_serves_total,
player_2_first_serve_points_won,
player_2_first_serve_points_total,
player_2_second_serve_points_won,
player_2_second_serve_points_total,
player_2_break_points_saved,
player_2_break_points_serve_total,
player_2_service_games_played,
player_2_return_rating,
player_2_first_serve_return_won,
player_2_first_serve_return_total,
player_2_second_serve_return_won,
player_2_second_serve_return_total,
player_2_break_points_converted,
player_2_break_points_return_total,
player_2_return_games_played,
player_2_service_points_won,
player_2_service_points_total,
player_2_return_points_won,
player_2_return_points_total,
player_2_under_pressure_rating,
player_2_win_rate]

for value1, value2 in zip(liste_1, liste_2):
    df[value1] = pd.DataFrame(value2)

# suppression colonnes winner et loser
col_to_drop = ['winner_B365', 'winner_PS', 'winner_name',
 'winner_rank',
 'winner_height_cm',
 'winner_weight_kg',
 'winner_handedness',
 'winner_backhand',
 'winner_at_home',
 'winner_age',
 'winner_nb_year_xp',
 'winner_match_duration',
 'winner_nb_sets',
 'winner_nb_games',
 'winner_nb_tiebreaks',
 'winner_nb_points',
 'winner_sets_won',
 'winner_games_won',
 'winner_tiebreaks_won',
 'winner_points_won',
 'winner_serve_rating',
 'winner_aces',
 'winner_double_faults',
 'winner_first_serves_in',
 'winner_first_serves_total',
 'winner_first_serve_points_won',
 'winner_first_serve_points_total',
 'winner_second_serve_points_won',
 'winner_second_serve_points_total',
 'winner_break_points_saved',
 'winner_break_points_serve_total',
 'winner_service_games_played',
 'winner_return_rating',
 'winner_first_serve_return_won',
 'winner_first_serve_return_total',
 'winner_second_serve_return_won',
 'winner_second_serve_return_total',
 'winner_break_points_converted',
 'winner_break_points_return_total',
 'winner_return_games_played',
 'winner_service_points_won',
 'winner_service_points_total',
 'winner_return_points_won',
 'winner_return_points_total',
 'winner_under_pressure_rating',
 'winner_win_rate',
               'loser_B365',
               'loser_PS',
 'loser_name',
 'loser_rank',
 'loser_height_cm',
 'loser_weight_kg',
 'loser_handedness',
 'loser_backhand',
 'loser_at_home',
 'loser_age',
 'loser_nb_year_xp',
 'loser_match_duration',
 'loser_nb_sets',
 'loser_nb_games',
 'loser_nb_tiebreaks',
 'loser_nb_points',
 'loser_sets_won',
 'loser_games_won',
 'loser_tiebreaks_won',
 'loser_points_won',
 'loser_serve_rating',
 'loser_aces',
 'loser_double_faults',
 'loser_first_serves_in',
 'loser_first_serves_total',
 'loser_first_serve_points_won',
 'loser_first_serve_points_total',
 'loser_second_serve_points_won',
 'loser_second_serve_points_total',
 'loser_break_points_saved',
 'loser_break_points_serve_total',
 'loser_service_games_played',
 'loser_return_rating',
 'loser_first_serve_return_won',
 'loser_first_serve_return_total',
 'loser_second_serve_return_won',
 'loser_second_serve_return_total',
 'loser_break_points_converted',
 'loser_break_points_return_total',
 'loser_return_games_played',
 'loser_service_points_won',
 'loser_service_points_total',
 'loser_return_points_won',
 'loser_return_points_total',
 'loser_under_pressure_rating',
 'loser_win_rate']

df.drop(col_to_drop, axis=1, inplace=True)

# Décision de suppression du nom du tournoi car trop de modalités
# et nous avons déjà at_home, Surface, Court, prize_money_euro
df.drop('Tournament', axis=1, inplace=True)

# Suppression de la Date car on a year, month et day
df.drop('Date', axis=1, inplace=True)

df = df[(df['player_1_B365'] != 0) & (df['player_2_B365'] != 0) & (df['player_1_PS'] != 0) & (df['player_2_PS'] != 0)]

df.to_csv('Sources/df_final_rolling_ready_30_with_bookodds.csv', index=False)

df
